Refactor cli into parts
