+++ proteus-cli/src/cli/args.rs
@@ -1,0 +1,127 @@
+use clap::{Arg, ArgAction, Command};
+
+pub fn build_cli() -> Command {
+    Command::new("Prot Play")
+        .version("1.0")
+        .author("Adam Howard <adam.thomas.howard@gmail.com>")
+        .about("Play Prot audio")
+        .arg(
+            Arg::new("seek")
+                .long("seek")
+                .short('s')
+                .value_name("TIME")
+                .help("Seek to the given time in seconds")
+                .conflicts_with_all(&["verify", "decode-only", "verify-only", "probe-only"]),
+        )
+        .arg(
+            Arg::new("GAIN")
+                .long("gain")
+                .short('g')
+                .value_name("GAIN")
+                .default_value("70")
+                .help("The playback gain"),
+        )
+        .arg(
+            Arg::new("impulse-response")
+                .long("ir")
+                .value_name("IMPULSE")
+                .help(
+                    "Impulse response path or attachment (e.g., file:ir.wav or attachment:ir.wav)",
+                ),
+        )
+        .arg(
+            Arg::new("bench-dsp")
+                .long("bench-dsp")
+                .action(ArgAction::SetTrue)
+                .help("Run a synthetic DSP benchmark and exit"),
+        )
+        .arg(
+            Arg::new("bench-sweep")
+                .long("bench-sweep")
+                .action(ArgAction::SetTrue)
+                .help("Run a sweep over multiple FFT sizes and exit"),
+        )
+        .arg(
+            Arg::new("bench-fft-size")
+                .long("bench-fft-size")
+                .value_name("SIZE")
+                .default_value("24576")
+                .help("FFT size for DSP benchmark"),
+        )
+        .arg(
+            Arg::new("bench-input-seconds")
+                .long("bench-input-seconds")
+                .value_name("SECONDS")
+                .default_value("1.0")
+                .help("Input length in seconds for DSP benchmark"),
+        )
+        .arg(
+            Arg::new("bench-ir-seconds")
+                .long("bench-ir-seconds")
+                .value_name("SECONDS")
+                .default_value("2.0")
+                .help("Impulse response length in seconds for DSP benchmark"),
+        )
+        .arg(
+            Arg::new("bench-iterations")
+                .long("bench-iterations")
+                .value_name("COUNT")
+                .default_value("5")
+                .help("Number of iterations for DSP benchmark"),
+        )
+        .arg(
+            Arg::new("start-buffer-ms")
+                .long("start-buffer-ms")
+                .value_name("MS")
+                .default_value("20")
+                .help("Amount of audio (ms) to buffer before starting playback"),
+        )
+        .arg(
+            Arg::new("decode-only")
+                .long("decode-only")
+                .help("Decode, but do not play the audio")
+                .conflicts_with_all(&["probe-only", "verify-only", "verify"]),
+        )
+        .arg(
+            Arg::new("probe-only")
+                .long("probe-only")
+                .help("Only probe the input for metadata")
+                .conflicts_with_all(&["decode-only", "verify-only"]),
+        )
+        .arg(
+            Arg::new("verify-only")
+                .long("verify-only")
+                .help("Verify the decoded audio is valid, but do not play the audio")
+                .conflicts_with_all(&["verify"]),
+        )
+        .arg(
+            Arg::new("verify")
+                .long("verify")
+                .short('v')
+                .help("Verify the decoded audio is valid during playback"),
+        )
+        .arg(
+            Arg::new("no-progress")
+                .long("no-progress")
+                .help("Do not display playback progress"),
+        )
+        .arg(
+            Arg::new("no-gapless")
+                .long("no-gapless")
+                .help("Disable gapless decoding and playback"),
+        )
+        .arg(
+            Arg::new("quiet")
+                .long("quiet")
+                .short('q')
+                .action(ArgAction::SetTrue)
+                .help("Suppress all console output"),
+        )
+        .arg(Arg::new("debug").short('d').help("Show debug output"))
+        .arg(
+            Arg::new("INPUT")
+                .help("The input file path, or - to use standard input")
+                .required_unless_present_any(["bench-dsp", "bench-sweep"])
+                .index(1),
+        )
+}
