+++ proteus-cli/src/cli/mod.rs
@@ -1,0 +1,2 @@
+pub mod args;
+pub mod bench;
